Create Shooting
So, for this to work, you'll need to create a prefab to instantiate as a projectile. I've called mine projectile, but the beauty of programming is you can name it "ceiling cat is watching you masturbate", if you want to really confuse hackers. You will also need to designate the first-person camera as main.